Fix: Passagem da class de teste para class de produção
