+++ rick_and_morty_app/lib/features/list_characters/controller/controller.dart
@@ -1,0 +1,1 @@
+export './list_characters_controller.dart';
