+++ rick_and_morty_app/lib/features/list_characters/controller/list_characters_controller.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/foundation.dart';
+
+import '../../../core/core.dart';
+import '../repository/repository.dart';
+
+class ListCharactersController {
+  IListCharactersRepository listCharactersRepository = ListCharactersRepository();
+  ListCharacter? listCharacter;
+
+  Future getListCharacters() async {
+    String data = await listCharactersRepository.getListChatactersRepository();
+    try {
+      listCharacter = listCharacterFromJson(data);
+    } catch (e) {
+      if (kDebugMode) print(e.toString());
+      throw DefalutStrings.unknownError;
+    }
+  }
+}
