+++ rick_and_morty_app/test/features/list_characters/controller/list_characters_controller_test.dart
@@ -1,27 +1,11 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:rick_and_morty_app/core/core.dart';
+import 'package:rick_and_morty_app/features/list_characters/controller/controller.dart';
 import 'package:rick_and_morty_app/features/list_characters/repository/list_characters_repository.dart';
 
 import '../repository/list_characters_repository_test.mocks.dart';
 import '../repository/mock_correct_response.dart';
-
-class ListCharactersController {
-  IListCharactersRepository listCharactersRepository = ListCharactersRepository();
-  ListCharacter? listCharacter;
-
-  Future getListCharacters() async {
-    String data = await listCharactersRepository.getListChatactersRepository();
-    try {
-      listCharacter = listCharacterFromJson(data);
-    } catch (e) {
-      if (kDebugMode) print(e.toString());
-      throw DefalutStrings.unknownError;
-    }
-  }
-}
 
 void main() {
   late ListCharactersController sut;
